Using clang on travis for faster builds, fixed class order for bindings.

+++ bindings/python/exporters_vec_types/__vec_exporter_11.cpp
@@ -15,6 +15,6 @@ using namespace fertilized;
 
 namespace pyfertilized {
   void __export_vectors_11() {
-    custom_vector_from_seq<typename fertilized::Forest<float,float,uint,std::vector<float>,std::vector<float>>::tree_ptr_t>();
+    custom_vector_from_seq<typename fertilized::Forest<int,int,uint,std::vector<float>,std::vector<float>>::tree_ptr_t>();
   };
 }  // namespace pyfertilized
